feat: fully replaced all commands with new structure
