+++ commands/react/edit.ts
@@ -1,7 +1,7 @@
 import {
+  ApplicationCommandOptionType,
   ChatInputCommandInteraction,
   parseEmoji,
-  PermissionsBitField,
 } from 'discord.js';
 import {
   GET_REACT_ROLE_BY_EMOJI,
@@ -9,38 +9,35 @@ import {
   UPDATE_REACT_ROLE_EMOJI_ID,
   UPDATE_REACT_ROLE_EMOJI_TAG,
 } from '../../src/database/queries/reactRole.query';
-import { Category } from '../../utilities/types/commands';
 import { RolePing } from '../../utilities/utilPings';
-import { SlashCommand } from '../slashCommand';
+import { SlashSubCommand } from '../command';
 
-export class ReactEditCommand extends SlashCommand {
-  constructor() {
-    super(
-      'react-edit',
-      'Edit any existing react roles emoji!',
-      Category.react,
-      [PermissionsBitField.Flags.ManageRoles]
-    );
-
-    this.addRoleOption(
-      'role',
-      'The role that belongs to the react role.',
-      true
-    );
-    this.addStringOption(
-      'new-emoji',
-      'The new emoji for the react role.',
-      true
-    );
+export class EditSubCommand extends SlashSubCommand {
+  constructor(baseCommand: string) {
+    super(baseCommand, 'edit', 'Edit any existing react roles emoji!', [
+      {
+        name: 'role',
+        description: 'The role that belongs to the react role.',
+        required: true,
+        type: ApplicationCommandOptionType.Role,
+      },
+      {
+        name: 'new-emoji',
+        description: 'The new emoji for the react role.',
+        required: true,
+        type: ApplicationCommandOptionType.String,
+      },
+    ]);
   }
 
   execute = async (interaction: ChatInputCommandInteraction) => {
     if (!interaction.guildId) {
-      return interaction.reply({
-        ephemeral: true,
-        content: `Hey! I don't see the guild ID anywhere... that's weird.`,
-      });
+      return this.log.debug('Missing guildID on interaction.');
     }
+
+    await interaction.deferReply({
+      ephemeral: true,
+    });
 
     const role = this.expect(interaction.options.getRole('role'), {
       message: `Somehow the role is missing! Please try again.`,
@@ -53,19 +50,17 @@ export class ReactEditCommand extends SlashCommand {
 
     const doesReactRoleExist = await GET_REACT_ROLE_BY_ROLE_ID(role.id);
     if (!doesReactRoleExist) {
-      return interaction.reply({
-        ephemeral: true,
-        content: `Hey! That role doesn't belong to an existing react role.`,
-      });
+      return interaction.editReply(
+        `Hey! That role doesn't belong to an existing react role.`
+      );
     }
 
     const parsedEmoji = parseEmoji(emoji);
 
     if (!parsedEmoji?.id && !parsedEmoji?.name) {
-      return interaction.reply({
-        ephemeral: true,
-        content: `Hey! I had an issue parsing whatever emoji you passed in. Please wait and try again.`,
-      });
+      return interaction.editReply(
+        `Hey! I had an issue parsing whatever emoji you passed in. Please wait and try again.`
+      );
     }
 
     const emojiContent = parsedEmoji.id ?? parsedEmoji.name;
@@ -76,12 +71,11 @@ export class ReactEditCommand extends SlashCommand {
     );
 
     if (doesEmojiBelongToReactRole) {
-      return interaction.reply({
-        ephemeral: true,
-        content: `Hey! That emoji belongs to a react role (${RolePing(
+      return interaction.editReply(
+        `Hey! That emoji belongs to a react role (${RolePing(
           doesEmojiBelongToReactRole.roleId
-        )})`,
-      });
+        )})`
+      );
     }
 
     const emojiTag = parsedEmoji?.id
@@ -98,11 +92,10 @@ export class ReactEditCommand extends SlashCommand {
       interaction.guildId
     );
 
-    return interaction.reply({
-      ephemeral: true,
-      content: `:tada: Successfully updated the react role (${
+    return interaction.editReply(
+      `:tada: Successfully updated the react role (${
         emojiTag ?? parsedEmoji.name
-      } - ${RolePing(role.id)}) :tada:`,
-    });
+      } - ${RolePing(role.id)}) :tada:`
+    );
   };
 }
